+++ platforms/opencl/src/OpenCLParallelKernels.h
@@ -76,7 +76,7 @@ private:
     class FinishComputationTask;
     OpenCLPlatform::PlatformData& data;
     std::vector<Kernel> kernels;
-    std::vector<long> completionTimes;
+    std::vector<long long> completionTimes;
     std::vector<int> contextTiles;
 };
